update error and sdk alert
