+++ static/app/views/replays/list/replaysErroneousDeadRageCards.tsx
@@ -1,8 +1,7 @@
-import {Fragment, ReactNode, useMemo} from 'react';
+import {ReactNode, useMemo} from 'react';
 import styled from '@emotion/styled';
 import {Location} from 'history';
 
-import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {Organization} from 'sentry/types';
 import EventView from 'sentry/utils/discover/eventView';
@@ -10,9 +9,6 @@ import useReplayList from 'sentry/utils/replays/hooks/useReplayList';
 import {useHaveSelectedProjectsSentAnyReplayEvents} from 'sentry/utils/replays/hooks/useReplayOnboarding';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
-import useProjectSdkNeedsUpdate from 'sentry/utils/useProjectSdkNeedsUpdate';
-import {EmptyStateSubheading} from 'sentry/views/replays/list/replaysList';
 import ReplayTable from 'sentry/views/replays/replayTable';
 import {ReplayColumn} from 'sentry/views/replays/replayTable/types';
 import {ReplayListLocationQuery} from 'sentry/views/replays/types';
@@ -115,18 +111,6 @@ function ReplaysErroneousDeadRageCards() {
     ReplayColumn.ACTIVITY,
   ];
 
-  const {
-    selection: {projects},
-  } = usePageFilters();
-
-  const MIN_DEAD_RAGE_CLICK_SDK = '7.60.1';
-
-  const needSDKUpgrade = useProjectSdkNeedsUpdate({
-    minVersion: MIN_DEAD_RAGE_CLICK_SDK,
-    organization,
-    projectId: projects.map(p => String(p)),
-  });
-
   return hasSessionReplay && !fetching && hasSentOneReplay ? (
     hasDeadRageCards ? (
       <SplitCardContainer>
@@ -141,19 +125,6 @@ function ReplaysErroneousDeadRageCards() {
           location={newLocation}
           organization={organization}
           visibleColumns={deadRageCols}
-          emptyMessage={
-            needSDKUpgrade.needsUpdate ? (
-              <Fragment>
-                {t('There are no items to display')}
-                <EmptyStateSubheading>
-                  {tct('[data] require an [sdkPrompt]', {
-                    data: <strong>Rage and dead clicks</strong>,
-                    sdkPrompt: <strong>{t('SDK version >= 7.60.1')}</strong>,
-                  })}
-                </EmptyStateSubheading>
-              </Fragment>
-            ) : undefined
-          }
         />
       </SplitCardContainer>
     ) : null
